+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+* Add a `/etc/portage/bashrc` to automatically signing kernel modules with `Portage`.
 * Now clean properly another disk used with option s`--separate-x DISK`.
 * For ZFS, use `blockdev --getpbsz` to find correct bloc (sector) size.
 * Populate `/etc/modules-load.d/` with modules found with `lsmod` (only few wifi's, flash usb related (ehci, ohci, xhci, etc)).